requirements.txt
Django==5.0.7
django-cors-headers==4.3.1
djangorestframework==3.15.2
inflection==0.5.1
psycopg2-binary==2.9.10
numpy==2.2.4
torch==2.6.0
matplotlib==3.10.1
pandas==2.2.3